follow and block mechanics

+++ db/entities/user_session.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from sqlalchemy import Integer, ForeignKey, String
+from sqlalchemy import BigInteger, ForeignKey, String
 
 from db.engine import Base
 
@@ -11,7 +11,7 @@ class UserSessionModel(Base):
     __tablename__ = 'user_sessions'
 
     session_id: Mapped[str] = mapped_column(String(43), primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.user_id'), nullable=False)
-    valid_until: Mapped[int] = mapped_column(Integer, nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('users.user_id'), nullable=False)
+    valid_until: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
     user = relationship('UserModel', back_populates='sessions', uselist=False)
